customer-service/requirements.txt
aio-pika==9.4.2
aioredis==2.0.1
aiormq==6.8.0
aiosqlite==0.20.0
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click-spinner==0.1.10
coverage==7.6.1
croniter==1.4.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fast-depends==2.4.7
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-utilities==0.2.0
faststream==0.5.16
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.10.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
packaging==24.1
pamqp==3.3.0
passlib==1.7.4
pika==1.3.2
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-mock==3.14.0
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
